requirements.txt
asgiref==3.11.1
coverage==7.13.5
dj-database-url==3.1.2
dj-static==0.0.6
Django==6.0.3
django-bootstrap3==26.1
django-stubs==6.0.0
django-stubs-ext==6.0.0
djangorestframework==3.17.1
factory_boy==3.3.3
Faker==40.11.1
gunicorn==25.1.0
iniconfig==2.3.0
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.1.0
pytest-django==4.12.0
sqlparse==0.5.5
static3==0.7.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
